refactor: Qt6 compatibility

+++ 3rdparty/QCodeEditor/src/QCodeEditor.cpp
@@ -536,7 +536,7 @@ void QCodeEditor::keyPressEvent(QKeyEvent* e) {
 
     // Shortcut for moving line to left
     if (m_replaceTab && e->key() == Qt::Key_Backtab) {
-      indentationLevel = std::min(indentationLevel, m_tabReplace.size());
+      indentationLevel = std::min(indentationLevel, (int)m_tabReplace.size());
 
       auto cursor = textCursor();
